jsonmanifest: split manifest.go into manifest.go and entry.go
